Feature/tree transformer and streamlit refactor

+++ cli/analyse_openings.py
@@ -6,40 +6,40 @@ from next_move.opening_directory import EcoDB
 
 # from next_move.visualiser.visualiser import Visualiser
 
+from pathlib import Path
 from tqdm import tqdm
 from typing import Optional
 
+from dotenv import load_dotenv, find_dotenv
+
 import click
+import os
+
+
+load_dotenv(find_dotenv())
 
 
 def run_analysis(player_id: str) -> Tree:
-    tree = Tree()
-    chess_com = ChessCom()
-    # visualiser = Visualiser()
-    games = chess_com.get_all_games(player_id)
-    stockfish = Stockfish("16/bin/stockfish")
-    eco_db = EcoDB("eco/openings.json")
+    if Path(f"next_move/cache/trees/{player_id}.json").exists():
+        return Tree.from_json(f"next_move/cache/trees/{player_id}.json")
+    else:
+        tree = Tree()
+        chess_com = ChessCom()
 
-    game_processor = GameProcessor(tree, stockfish, eco_db, player_id)
+        games = chess_com.get_all_games(player_id)
+        stockfish = Stockfish(os.environ["STOCKFISH_DIR"])
+        eco_db = EcoDB("eco/openings.json")
 
-    for game in tqdm(games):
-        game_processor.process_game(game)
+        game_processor = GameProcessor(tree, stockfish, eco_db, player_id)
 
-    stockfish.quit()  # it's fine for now, but will need to refactor this into a context manager
+        for game in tqdm(games):
+            game_processor.process_game(game)
 
-    # visualiser.sankey(
-    #     **game_processor.tree.to_sankey(prune_below_count=5), path="tree.html"
-    # )
-    # visualiser.timeline(
-    #     game_processor.tree.to_timeline(breakdown="M"), path="timeline.png"
-    # )
+        stockfish.quit()  # it's fine for now, but will need to refactor this into a context manager
 
-    # df = tree.to_opening_strength()
-    # df.to_csv("opening_strength.csv")
+        tree.to_json(f"next_move/cache/trees/{player_id}.json")
 
-    tree.to_json(f"next_move/cache/trees/{player_id}.json")
-
-    return tree
+        return tree
 
 
 @click.command()
